requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
blinker==1.9.0
cachetools==5.5.2
catkin-pkg==1.0.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.2.1
colcon-cmake==0.2.29
colcon-common-extensions==0.3.0
colcon-core==0.19.0
colcon-defaults==0.2.9
colcon-devtools==0.3.0
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.9
colcon-recursive-crawl==0.2.3
colcon-ros==0.5.0
colcon-test-result==0.3.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
coverage==7.9.2
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
distlib==0.4.0
distro==1.9.0
docutils==0.21.2
efficientnet_pytorch==0.7.1
empy==4.2
executing==2.2.1
filelock==3.25.2
firebase==4.0.1
Flask==3.0.0
Flask-Cors==4.0.0
fonttools==4.62.1
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huepy==1.2.1
humanfriendly==10.0
id==1.5.0
idna==3.11
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keyring==25.6.0
kiwisolver==1.5.0
livereload==2.7.1
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
markdown2==2.5.4
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
nh3==0.3.0
npm==0.1.1
numpy==2.4.3
oauthlib==3.3.1
openai==2.5.0
opencv-python==4.13.0.92
optional-django==0.1.0
packaging==26.0
pandas==2.3.2
parso==0.8.5
pdoc3==0.11.6
pillow==12.1.1
platformdirs==4.4.0
pluggy==1.6.0
polars==1.39.2
polars-runtime-32==1.39.2
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.2.2
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.3.2
pyreadline3==3.5.4
pytest==8.4.1
pytest-cov==6.2.1
pytest-repeat==0.9.4
pytest-rerunfailures==15.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
pyzmq==27.0.2
readme_renderer==44.0
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.17.1
seaborn==0.13.2
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
source==1.2.0
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.10.0
torchaudio==2.8.0
torchvision==0.25.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
twine==6.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.4.23
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.6.3
vcstool==0.3.0
vermin==1.6.0
vite==1.5.2
wcwidth==0.2.13
Werkzeug==3.1.3
